Refactor components into events/description/ and fix details box styling
