+++ components/core/divider/CoreDivider.tsx
@@ -5,8 +5,7 @@ import { Divider } from "@material-ui/core";
 type Props = React.ComponentProps<typeof Divider>;
 
 const CoreDivider: React.FC<Props> = (props: Props) => {
-  const { ...rest } = props;
-  return <Divider {...rest} />;
+  return <Divider {...props} />;
 };
 
 export default CoreDivider;
