+++ pages/events/[id].tsx
@@ -9,7 +9,7 @@ import {
 import ErrorPage from "next/error";
 import { useRouter } from "next/router";
 
-import EventsPageDescription from "components/events/EventsPageDescription";
+import EventsPageDescription from "components/events/description/EventsPageEventDescription";
 import config from "config";
 import { getAllEventIds, getEventById } from "server/actions/events";
 import { getNonprofitById } from "server/actions/nonprofit";
